+++ Q3/Q3.h
@@ -3,8 +3,8 @@
 
 
 void *createCodec(char key[62]);
-int encode(char *textin, char *textout, int len, void *codec);
-int decode(char *textin, char *textout, int len, void *codec);
+int encode(char *textin, char *textout, long len, void *codec);
+int decode(char *textin, char *textout, long len, void *codec);
 void freeCodec(void *codec);
 
 #endif
